+++ view/src/net/katsuster/strview/media/mkv/EBMLlacing.java
@@ -79,8 +79,8 @@ public class EBMLlacing extends EBMLvint
             throws CloneNotSupportedException {
         EBMLlacing obj = (EBMLlacing)super.clone();
 
-        obj.vint_head = vint_head.clone();
-        obj.vint_val = vint_val.clone();
+        obj.vint_head = (UInt)vint_head.clone();
+        obj.vint_val = (UInt)vint_val.clone();
 
         return obj;
     }
